pit: fix GPT start and end LBA calculations
This patch is for fixing pit and GPT lba addressing
because there was no consideration
about GPT backup header space.

And I think a definition of PIT_PART_META is a little bit
confusing. So now this means total space for GPT
at the head of storage. So you can guess PIT start LBA is
PIT_PART_META.

Change-Id: Idf14a5d7b4452bd8a98fe302dc4594756c5af82f
Signed-off-by: khyun.cho <khyun.cho@samsung.com>