ufs: fix the LU configuration method
1. To calculate a value of dNumAllocUnits from target Unit descriptor,
a value of wEnhanced1CapAdjFac from geometry dectriptor was
referred without endian swapping even if it's more than one bytes.
This might make unexpected results. So I fixed to swap a value of
wEnhanced1CapAdjFac before calculating the target capacity.

2. A target value of dNumAllocUnits of LU #0 is calculated by subtracking
from the device's total capacity to allocated capacities of all other
LUs, but the former's unit was used as a number of 'segments' and
the latter's unit was used as a number of 'allocation units'.
So I replaced the former's unit with allocation unit.

Change-Id: Idb4350f71333968ed734042ee9e39875dde4d38d
Signed-off-by: Kiwoong Kim <kwmad.kim@samsung.com>