fastboot: support 'partinfo:', 'diskinfo:' and 'diskdump:'
This is to enable to read back storage.
Host system submits 'partinfo:' with partition name
or 'diskinfo:' with block range. And then booloader
returns available BLKRANGE. If the values are zero,
that means impossible to read anywhere with the input given before.
After that, host system should terminate read back function.

If the value are non-zero, host system submits 'diskdump:'
and bootloader does read back function, transfer data to host system
and returns BLKCOUNT as the result with actual transffered size.

Change-Id: I844bbf64c4a82a5881f0d9f8646b11ec609b1722
Signed-off-by: Kiwoong Kim <kwmad.kim@samsung.com>