Sam: Don't try to boot boot.img ver 3+
Current bootloader is only capable of loading and running Android boot
image version 2. In case when boot.img ver 3 is flashed, attempt to load
it leads to data abort, thus making the board bricked (fastboot isn't
functional in that case). In order to avoid that, check boot.img header
version first, and if it's greater than 2, print corresponding error and
fallback to fastboot mode gracefully, so that the user is able to
re-flash correct images.

There two functions have to be called before going to fastboot:

    sdm_encrypt_secdram();
    dfd_set_dump_en_for_cacheop(0);

because otherwise the consequent reboot crashes with data abort in MMC
driver.

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>