[FIXES] exynos3830: Change kernel loading address to 0x80000000
Mainline kernel now requires to be loaded to 2 MiB aligned address,
showing this warning:

    [Firmware Bug]: Kernel image misaligned at boot, please fix your bootloader!

Earlier Linux kernel required to leave 512 KiB at the RAM beginning for
swapper page tables (BSS section). Now it's not needed anymore [1], and
TEXT_OFFSET was set to 0x0. On the other hand, it's now required that
bootloader be aligned by 2 MiB [2]. That said, let's set the loading
address of kernel image to 0x80000000 to fix mentioned warning.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=cfa7ede20f133cc81cef01dc3a516dda3a9721ee
[2] https://www.spinics.net/lists/arm-kernel/msg813171.html

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>