[FIXES] exynos3830: Relocate BL2 tag to safe location
We want to use 0x80000000 as a KERNEL_BASE (to load kernel image to).
But this location (DRAM_BASE) is already used by BL2 -- it writes
0xabcdef value here to indicate it was executed. Let's relocate that tag
value to more safe location (0xff000000) at early platform init, and check
that relocated address later instead of checking 0x80000000.

The 0xff000000 location was chosen because it should be vacant: it resides
upper from bootloader:
  - bootloader image (+stack and BSS): 0xf8800000..f9000000
  - RAM is mapped up to 0xffffffff

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>