[E850-96] Implement a way to mount rootfs from super partition
In case when upstream dtb is detected and ramdisk is not present in
"boot" partition, don't add linux,initrd* properties to the /chosen
node. Instead add "root=" param to kernel cmdline, so that rootfs is
mounted from "super" partition. It allows one to run regular Linux
rootfs (like OpenEmbedded rootfs or Debian rootfs) while having the same
Android eMMC partition table. It also leaves the option to use rootfs
from the ramdisk, if it's present in "boot" partition.

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>