[FIXES] exynos3830: fdt: Handle invalid value in set_fdt_val()
set_fdt_val() expects 'value' param in a form of a number, e.g.
"<0xabcdef00>". In case it's not a number, like "<null>", set_fdt_val()
stucks in the infinite loop, which leads to data abort eventually. Let's
check if 'value' param is actually a number, before trying to parse it
further. If it's not -- show the error message and return some error
code.

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>